% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{app}
\alias{app}
\title{Launch shiny app}
\usage{
app()
}
\value{
Launch in web browser the app
}
\description{
Launch the application for mapping and evaluating the IUCN status
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
