% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs_theme_conr.R
\name{bs_theme_conr}
\alias{bs_theme_conr}
\title{Bootstrap theme for ConR}
\usage{
bs_theme_conr()
}
\value{
A \code{\link[bslib:bs_theme]{bslib:: bs_theme()}} object.
}
\description{
Bootstrap theme for ConR
}
