% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig.R
\name{collect_rast}
\alias{collect_rast}
\title{Extract a raster}
\usage{
collect_rast(names, XY_sf = NULL)
}
\arguments{
\item{names}{string name of the table}

\item{XY_sf}{sf object of dataset}
}
\value{
A sf object
}
\description{
Extract a raster from the database
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
