% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig.R
\name{conn_mydb_shp}
\alias{conn_mydb_shp}
\title{Load the database}
\usage{
conn_mydb_shp(pass = NULL, user = NULL)
}
\arguments{
\item{pass}{string}

\item{user}{string}
}
\value{
The database is loaded
}
\description{
Load the SIG database and ask for password
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
