% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conr-server.R
\name{conr_server}
\alias{conr_server}
\title{Application's server}
\usage{
conr_server()
}
\value{
No value.
}
\description{
Application's server
}
\seealso{
\itemize{
\item \code{\link[=conr_ui]{conr_ui()}} for UI part.
\item \code{\link[=launch]{launch()}} to launch application.
}
}
