% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conrappli-package.R
\name{conrappli-exports}
\alias{conrappli-exports}
\alias{\%>\%}
\title{conrappli exported operators and S3 methods}
\description{
The following functions are imported and then re-exported
from the apexcharter package to avoid listing packages
as Depends of conrappli
}
