% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conrappli-package.R
\docType{package}
\name{conrappli-package}
\alias{conrappli}
\alias{conrappli-package}
\title{Web Interface for 'ConR' package}
\description{
Shiny application for mapping and computing several range parameters for plant species
Dedicated to Central Africa. Provide preliminary IUCN evaluation following Criterion B.
Based on \href{https://github.com/gdauby/ConR}{ConR} package computation.
}
\author{
Gilles Dauby
}
