% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_accuracy.R
\name{coord_accuracy}
\alias{coord_accuracy}
\title{compute accuracy of coordinates}
\usage{
coord_accuracy(.data, col_x, col_y, rounding = TRUE)
}
\arguments{
\item{.data}{A \code{data.frame} with longitude and latitude in decimal degrees}

\item{col_x}{Column name containing the longitude.}

\item{col_y}{Column name containing the latitude.}

\item{rounding}{Round or not secondes.}
}
\value{
tibble provided with an additional column \code{calc_accuracy}.
}
\description{
convert into degrees minutes seconds and give code from 1 (low accuracy) to 8 (high accuracy) of accuracy
}
\examples{
test <- dplyr::tibble(x = c(8.8633, 6.2, 4), y = c(0.1321, 0.22, 1))
coord_accuracy(test, col_x = "x", col_y = "y")


}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
