% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig.R
\name{extract_overlap_shp}
\alias{extract_overlap_shp}
\title{Evaluate intersection}
\usage{
extract_overlap_shp(XY, col_x = "ddlon", col_y = "ddlat")
}
\arguments{
\item{XY}{data.frame with two columns with geographical coordinates (latitude, longitude)}

\item{col_x}{string column name indicating longitude}

\item{col_y}{string column name indicating latitude}
}
\value{
A list with two elements
}
\description{
Evaluate the intersection between coordinates and bounding boxes of the database
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
