% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering_out.R
\name{filtering_out}
\alias{filtering_out}
\alias{filtering_out_data}
\title{Filtering out}
\usage{
filtering_out_data(data, threshold)
}
\arguments{
\item{data}{A \code{data.frame} or similar.}

\item{threshold}{an integer}
}
\value{
\itemize{
\item \code{data_filtered} \code{data.frame} same as data filtered out according to threshold
\item \code{validation_rules} a set of validation rules defined with \code{\link[validate:validator]{validate::validator()}}.
}
}
\description{
Filertting out based on numner of nunique occurences
}
