% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hansen_deforestation_aggreg}
\alias{hansen_deforestation_aggreg}
\title{Deforestation estimates for Central Africa}
\format{
raster object
}
\usage{
hansen_deforestation_aggreg
}
\description{
Raster of 0.1 decimal degrees resolution indicating proportion of cells 25m from Hansen where deforestation occur since 2001
}
\keyword{datasets}
