% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.R
\name{launch}
\alias{launch}
\title{Launch ConR application}
\usage{
launch(lang = c("en", "fr"))
}
\arguments{
\item{lang}{Language to be used in the application : english ("en") or french ("fr").}
}
\value{
A \code{\link[shiny:shinyApp]{shiny::shinyApp()}} object.
}
\description{
Launch ConR application
}
\examples{
if (interactive()) {

  # A Launch application in browser
  launch()

}
}
