% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{logging_concessions_CA}
\alias{logging_concessions_CA}
\title{Logging concessions Central Africa}
\format{
sf object
}
\usage{
logging_concessions_CA
}
\description{
sf multypolygon
}
\keyword{datasets}
