% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\docType{data}
\name{mineral_deposit}
\alias{mineral_deposit}
\title{Mineral deposits in Africa with a 10 km buffer}
\format{sf object}
\usage{
mineral_deposit
}
\description{
sf multypolygon of mineral deposits either exploited or not with a 10 km buffer
}
\keyword{datasets}
