% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-criterion_b.R
\name{module-analysis}
\alias{module-analysis}
\alias{criterion_b_ui}
\alias{criterion_b_server}
\title{Criterion B analysis}
\usage{
criterion_b_ui(id)

criterion_b_server(
  id,
  data_r = reactive(NULL),
  threat_sig_r = reactive(NULL),
  taxa_selected_r = reactive(NULL),
  table_overlap_r = reactive(NULL)
)
}
\arguments{
\item{id}{Module's ID.}

\item{data_r}{A \code{reactive} function returning a \code{data.frame}.}

\item{threat_sig_r}{A \code{reactive} function returning spatial data to use in analysis.}

\item{taxa_selected_r}{A \code{reactive} function returning the taxa to select by default.}
}
\value{
\itemize{
\item UI: HTML tags that can be included in the UI part of the application.
\item Server: a \code{\link[shiny:reactive]{shiny::reactive()}} function returning a \code{data.frame}.
}
}
\description{
Criterion B analysis
}
