% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-data-display.R
\name{module-data-display}
\alias{module-data-display}
\alias{data_display_ui}
\alias{data_display_server}
\title{Display data Module}
\usage{
data_display_ui(id)

data_display_server(id, data_r = reactive(NULL))
}
\arguments{
\item{id}{Module's ID.}

\item{data_r}{A \code{reactive} function returning a \code{data.frame}.}
}
\value{
\itemize{
\item UI: HTML tags that can be included in the UI part of the application.
\item Server: No value.
}
}
\description{
Display data Module
}
