% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-data-map.R
\name{module-data-map}
\alias{module-data-map}
\alias{data_map_ui}
\alias{data_map_server}
\title{Map point selection Module}
\usage{
data_map_ui(id)

data_map_server(id, data_r = reactive(NULL))
}
\arguments{
\item{id}{Module's ID.}

\item{data_r}{A \code{reactive} function returning a \code{data.frame}.}
}
\value{
\itemize{
\item UI: HTML tags that can be included in the UI part of the application.
\item Server: a \code{\link[shiny:reactive]{shiny::reactive()}} function returning a \code{list}.
}
}
\description{
Map point selection Module
}
