% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-data-import-polygon.R
\name{module-data-polygon}
\alias{module-data-polygon}
\alias{data_import_polygon_ui}
\alias{data_import_polygon_server}
\title{Import data from Rainbio via polygon drawing Module}
\usage{
data_import_polygon_ui(id)

data_import_polygon_server(id)
}
\arguments{
\item{id}{Module's ID.}
}
\value{
\itemize{
\item UI: HTML tags that can be included in the UI part of the application.
\item Server: a \code{\link[shiny:reactive]{shiny::reactive()}} function returning a \code{data.frame}.
}
}
\description{
Import data from Rainbio via polygon drawing Module
}
