% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-data-import-rainbio.R
\name{module-data-rainbio}
\alias{module-data-rainbio}
\alias{data_import_rainbio_ui}
\alias{data_import_rainbio_server}
\title{Import data from Rainbio Module}
\usage{
data_import_rainbio_ui(id, from = c("file", "copypaste"))

data_import_rainbio_server(id)
}
\arguments{
\item{id}{Module's ID.}

\item{from}{Type of data input.}
}
\value{
\itemize{
\item UI: HTML tags that can be included in the UI part of the application.
\item Server: a \code{\link[shiny:reactive]{shiny::reactive()}} function returning a \code{data.frame}.
}
}
\description{
Import data from Rainbio Module
}
