% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-data-source.R
\name{module-data-source}
\alias{module-data-source}
\alias{data_source_ui}
\alias{data_source_server}
\title{Data source Module}
\usage{
data_source_ui(id)

data_source_server(id)
}
\arguments{
\item{id}{Module's ID.}

\item{data_r}{A \code{reactive} function returning a \code{data.frame}.}
}
\value{
\itemize{
\item UI: HTML tags that can be included in the UI part of the application.
\item Server: a \code{\link[shiny:reactive]{shiny::reactive()}} function returning a \code{data.frame}.
}
}
\description{
Data source Module
}
