% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-data-validation.R
\name{module-data-validation}
\alias{module-data-validation}
\alias{data_validation_ui}
\alias{data_validation_server}
\title{Data Validation Module}
\usage{
data_validation_ui(id)

data_validation_server(id, data_r = reactive(NULL))
}
\arguments{
\item{id}{Module's ID.}

\item{data_r}{A \code{reactive} function returning a \code{data.frame}.}
}
\value{
\itemize{
\item UI: HTML tags that can be included in the UI part of the application.
\item Server: a \code{\link[shiny:reactive]{shiny::reactive()}} function returning a \code{data.frame}.
}
}
\description{
Data Validation Module
}
