% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-data.R
\name{module-data}
\alias{module-data}
\alias{data_ui}
\alias{data_server}
\title{Data Module}
\usage{
data_ui(id)

data_server(id)
}
\arguments{
\item{id}{Module's ID.}
}
\value{
\itemize{
\item UI: HTML tags that can be included in the UI part of the application.
\item Server: a \code{\link[shiny:reactive]{shiny::reactive()}} function returning a \code{data.frame}.
}
}
\description{
This module allow to import data into the application,
to map input's columns to variable of interest for the simulation
and to validate that all is in order to proceed.
}
