% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-mapping.R
\name{module-mapping}
\alias{module-mapping}
\alias{mapping_ui}
\alias{mapping_server}
\title{Map point selection Module}
\usage{
mapping_ui(id)

mapping_server(id, data_r = reactive(NULL))
}
\arguments{
\item{id}{Module's ID.}

\item{data_r}{A \code{reactive} function returning a \code{data.frame}.}
}
\value{
\itemize{
\item UI: HTML tags that can be included in the UI part of the application.
\item Server: a \code{\link[shiny:reactive]{shiny::reactive()}} function returning a \code{list}.
}
}
\description{
Map point selection Module
}
