% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-report.R
\name{module-report}
\alias{module-report}
\alias{summary_report_ui}
\alias{summary_report_server}
\title{Summary report}
\usage{
summary_report_ui(id)

summary_report_server(
  id,
  data_r = reactive(NULL),
  results_r = reactive(NULL),
  data_sf_r = reactive(NULL),
  polygon_r = reactive(NULL)
)
}
\arguments{
\item{id}{Module's ID.}
}
\value{
\itemize{
\item UI: HTML tags that can be included in the UI part of the application.
\item Server: a \code{\link[shiny:reactive]{shiny::reactive()}} function returning a \code{data.frame}.
}
}
\description{
Summary report
}
