% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\docType{data}
\name{protected_areas}
\alias{protected_areas}
\title{Protected areas of Africa}
\format{sf object}
\usage{
protected_areas
}
\description{
An sf object of protected areas in Africa
}
\keyword{datasets}
