% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbif.R
\name{query_gbif_poly}
\alias{query_gbif_poly}
\title{Retrieve GBIF occurrences from polygon}
\usage{
query_gbif_poly(poly, threshold = 500)
}
\arguments{
\item{poly}{A sf}
}
\value{
A \code{tibble}.
}
\description{
Retrieve GBIF occurrences from polygon
}
