% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainbio.R
\name{query_rb_taxa}
\alias{query_rb_taxa}
\title{Extract the Rainbio database from a species list}
\usage{
query_rb_taxa(species = NULL, idtax = NULL, only_checked_georef = TRUE)
}
\arguments{
\item{species}{Names of species to search for.}

\item{idtax}{idtax.}

\item{only_checked_georef}{logical, whether filtering out occurrences with non validated georeferencing.}
}
\value{
A list with the sf of the Rainbio database extracted, the polygon used to extract, a tibble with idtax
}
\description{
Extract from the Rainbio database of all records from a species list
}
\examples{
\dontrun{
query_rb_taxa(species = c("diospyros iturensis", "anthonotha macrophylla"))
}
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
