% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\docType{data}
\name{rast_mayaux}
\alias{rast_mayaux}
\title{Raster or impacted land cover proportion}
\format{raster}
\usage{
rast_mayaux
}
\description{
Raster of land cover proportion impacted by human, from Mayaux et al. 2003
}
\keyword{datasets}
