% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbif.R
\name{retrieve_occ_data}
\alias{retrieve_occ_data}
\title{Retrieve GBIF occurrences}
\usage{
retrieve_occ_data(specieskey)
}
\arguments{
\item{specieskey}{A vector of species key.}
}
\value{
A \code{tibble}.
}
\description{
Retrieve GBIF occurrences
}
\examples{
\dontrun{

infos <- search_species_info(c("Uapaca niangadoumae"))
 retrieve_occ_data(infos$specieskey)

}
}
