% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbif.R
\name{search_species_info}
\alias{search_species_info}
\title{Search species information}
\usage{
search_species_info(
  species_name,
  match_type = c("exact", "confidence"),
  confidence_level = 95
)
}
\arguments{
\item{species_name}{Character of species names.}

\item{match_type}{Type of match for scientific names found: exact match or above confidence level specified.}

\item{confidence_level}{Confidence level to use for matching scientific names.}
}
\value{
A \code{tibble} with info found for each species provided.
}
\description{
Search species information
}
\examples{
\dontrun{
search_species_info(c("Uapaca niangadoumae"))
search_species_info(c("Uapaca niangadoumae"), match_type = "exact")
search_species_info(c("Uapaca niangadoumae", "do not exist"))
}
}
