% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validation}
\alias{validation}
\alias{prepare_data_validate}
\alias{validation_rules}
\alias{exclude_violating_records}
\alias{identify_violating_records}
\alias{extract_violating_records}
\title{Validation of data}
\usage{
prepare_data_validate(
  data,
  lat = "Latitude",
  lon = "Longitude",
  sci_names = "Taxa"
)

validation_rules()

exclude_violating_records(data)

identify_violating_records(data)

extract_violating_records(data)
}
\arguments{
\item{data}{A \code{data.frame} or similar.}

\item{lat, lon, sci_names}{Variables names used for validation.}
}
\value{
\itemize{
\item \code{prepare_data_validate()} \code{data.frame} with new columns.
\item \code{validation_rules} a set of validation rules defined with \code{\link[validate:validator]{validate::validator()}}.
}
}
\description{
Those functions are used to validate data imported by user:
\itemize{
\item \code{prepare_data_validate()} create new validation columns for relevant criteria.
\item \code{validation_rules} return validation rules.
}
}
