% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig.R
\name{collect_shp}
\alias{collect_shp}
\title{Extract a shapefile}
\usage{
collect_shp(table_names, XY_sf = NULL)
}
\arguments{
\item{XY_sf}{sf object of dataset}

\item{names}{string name of the table}
}
\value{
A sf object
}
\description{
Extract a shapefile from the database
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
