% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conr-ui.R
\name{conr_ui}
\alias{conr_ui}
\title{UI - ConR user interface}
\usage{
conr_ui(lang = c("en", "fr"))
}
\arguments{
\item{lang}{Language to be used in the application : english ("en") or french ("fr").}
}
\value{
An UI definition.
}
\description{
UI - ConR user interface
}
\seealso{
\itemize{
\item \code{\link[=conr_server]{conr_server()}} for server part.
\item \code{\link[=launch]{launch()}} to launch application.
}
}
