% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.R
\name{launch}
\alias{launch}
\title{Launch ConR application}
\usage{
launch()
}
\value{
A \code{\link[shiny:shinyApp]{shiny::shinyApp()}} object.
}
\description{
Launch ConR application
}
\examples{
if (interactive()) {

  # A Launch application in browser
  launch()

}
}
