% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-data-import-gbif.R,
%   R/module-data-import.R
\name{data_import_gbif_server}
\alias{data_import_gbif_server}
\alias{module-data-import}
\alias{data_import_ui}
\alias{data_import_server}
\title{Import data Module}
\usage{
data_import_gbif_server(id)

data_import_ui(id)

data_import_server(id)
}
\arguments{
\item{id}{Module's ID.}
}
\value{
\itemize{
\item UI: HTML tags that can be included in the UI part of the application.
\item Server: a \code{\link[shiny:reactive]{shiny::reactive()}} function returning a \code{data.frame}.
}
}
\description{
Import data Module
}
