% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-mapping.R
\name{module-mapping}
\alias{module-mapping}
\alias{mapping_ui}
\alias{mapping_server}
\title{Map point selection Module}
\usage{
mapping_ui(id)

mapping_server(
  id,
  data_r = reactive(NULL),
  data_latlon_r = reactive(NULL),
  trigger_map_r = reactive(FALSE),
  main_session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{id}{Module's ID.}

\item{data_r}{A \code{reactive} function returning a \code{data.frame}.}

\item{data_latlon_r}{A \code{reactive} function returning a \code{data.frame} with coordinate only.}

\item{trigger_map_r}{A \code{reactive} function to trigger map update.}

\item{main_session}{The main Shiny session to navigate between tabs.}
}
\value{
\itemize{
\item UI: HTML tags that can be included in the UI part of the application.
\item Server: a \code{\link[shiny:reactive]{shiny::reactive()}} function returning a \code{list}.
}
}
\description{
Map point selection Module
}
